api/requirements.txt
fastapi==0.115.0
python-multipart==0.0.18
markdown==3.7
pygments==2.18.0
premailer==3.10.0
pillow==11.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.0
